Add proper initialization of Event variables
ZFinderEvent internal variables are now set to obviously wrong values
(Z mass = -1 for example) on creation so that failed events are easy to
screen out. The electron pointers are initialized to NULL as well.

ZFinderPlotter now checks to make sure an event is good before plotting
it, preventing segfaults from reading from NULL pointers.

zfinder_cfg.py warns if set to not run on all events.