Fix a bug in the parsing of free-form '<' cuts
Instead of returning the position of the '<' on the line for slicing,
the code looked at the position of the '>'. Since this doesn't exist on
the line (or shouldn't ever), the slice point was the end of the string,
and hence a "" was converting to the double 0.0. This caused any cut
with a '<' to remove all events with a positive value in the selected
variable.