Fix trigger electron setting bug in ZFinderEvent
Trigger electrons were always being set equal to the actual reco
electrons due to shadowing a class variable. This was caught in the
first version of the code, but not all instances were changed
(unfortunately including the most important one, where the electron is
actually set).

Also fix the print shortcuts, and add one for trigger electrons.