Fix bug that skipped all Denominator events
Also update handling of degeneracy. We now have a three-way state: No
degeneracy, degeneracy in the denominator, and degeneracy in the
denominator and numerator.

Additionally we added a constructor to CutLevel to initialize all values
to false.

We also catch ZDefs with fewer than two levels of cuts and reject them.