Change acc(ALL) and add acc(DETECTOR)
acc(ALL) now includes all events, even those that would not hit part
of the detector. The intent is to use this to track the number of events
in a sample, which is useful when normalizing MC by its cross section
and event count. This is implemented by *HARDCODING* it to be true.

acc(DETECTOR) replaces the old version of acc(ALL). It is true if any
other acc() selection is true, excluding the new definition of acc(ALL).

To summarize:
    * acc(ALL): Always true.
    * acc(DETECTOR): True if the particle is somewhere in the detector.