Add efficiency passing to ZDefinitionPlotter
We take the efficiency from cutlevel. Since we don't know if the passing
set is Tag 0 Probe 1, or Tag 1 Probe 0, we take t0p1 if it passes,
otherwise we take t1p0 if it passes. If neither passes we use 1.

This might cause a *BIAS* if t0p1 and t1p0 both pass, but are slightly
different.