Add Gen Mass cut for MC ZDefinitions
It can now be specified in the ZDeffintion PSets if the mass cut should
be madd on the Reco mass or the Truth mass.

To help indicate what is being done to the consume of the output data,
the mass cut level name is changed depending on what quantity was used
as follows:
    * Reco:  "... < M_{ee} < ..."
    * Truth: "... < GEN M_{ee} < ..."

Also updated how we make ZDefinitions. Previously we made one set of
ZDefinitions and used these to make two sets of Plots and Workspaces,
one for Truth and one for Reco. We now make two sets of ZDefs, one for
Truth and one for Reco, and use each set to make the corresponding set
of plots.