Update ZDefinitions_cfi with Truth Mass cut option
Allow the user to specify that the mass cut should be made on the Truth
mass.

Added a set of cuts that mirror just the kinematic cuts from our
combined results, in both Generator quantities and Reco.

Also added a set of ZDefinitions without trigger requirements (mainly
for MC without emulated triggers).