Add pileup weighting scripts
pileupCalc.py is from CMSSW, but it is convenient to have in our
repository as well.

histo_to_vector.py prints out the bins in a histogram, like the ones
produced by pileupCalc.py, so that they can be turned into a vector that
is used in the reweighting code.