Split the configuration files into data/mc
There are now two configuration files:
    zfinder_data_cfg.py for data
    zfinder_mc_cfg.py for mc

The main difference is that the file to run on data includes a Lumi JSON
to restrict running to only good lumisegments. Additionally, each has
their own zdefinitions (the mc version does not have trigger matching
turned on).

To do this, we have split the configuration files into different cfi's,
and have imported these. This avoids code duplication.

Additionally we now read in the Lumi JSON file directly, instead of
converting it with a script and reading that. This makes it so that
there is only one file controlling what lumisegments are run on, instead
of two.