Update ZFinderEvent to allow 0,1 Reco Electrons
In order to do this, the following changes have been made:
    * If there is only one Reco Electron, it is set to e0
    * Events are required to have two Reco electrons _OR_ two Truth
      electrons in order to advance
    * Various "electron != NULL" tests were moved, removed, or split to
      allow events with fewer than 2 electrons to survive.

These 0,1 electron events are _STILL_ removed by the cut stages because
they do not have a second electron to pass the second set of cuts, and
are (at that point) marked as having 0 efficiency. They are not, in
general, saved to RooWorkspaces, although this might be changed in the
future.