Remove FollowElectron, add GetNakedElectron
FollowElectron was passed a pointer to a pointer and moved that pointer
to point to the naked electron, while returning a deep copy of the born
electron the pointers originally pointed to.

GetNakedElectron instead takes a pointer to a born electron, and returns
a pointer to the naked electron. It does not change the input born
electron at all, and makes no copies.