Fix bug in fabs(ETA) bug in EfficiencyTable
Eta symmetry is assumed in the construction of the scale factor tables
that are stored in EfficiencyTable, but EfficiencyTable was not using
fabs(eta) when looking up efficiencies. This lead to all lookups of
points with a negative value of eta failing, and hence leaving their
scale factors set to 1.

fabs(eta) is now used, and a warning is printed if a row with negative
eta values is inserted into the table.